# web/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/provider-configs": {
      "get": {
        "tags": [
          "provider-configs"
        ],
        "summary": "List Provider Configs",
        "operationId": "list_provider_configs_api_v1_provider_configs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderConfigResponse"
                  },
                  "type": "array",
                  "title": "Response List Provider Configs Api V1 Provider Configs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "provider-configs"
        ],
        "summary": "Create Provider Config",
        "operationId": "create_provider_config_api_v1_provider_configs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provider-configs/{provider_id}": {
      "patch": {
        "tags": [
          "provider-configs"
        ],
        "summary": "Update Provider Config",
        "operationId": "update_provider_config_api_v1_provider_configs__provider_id__patch",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "provider-configs"
        ],
        "summary": "Delete Provider Config",
        "operationId": "delete_provider_config_api_v1_provider_configs__provider_id__delete",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provider-configs/{provider_id}/test": {
      "post": {
        "tags": [
          "provider-configs"
        ],
        "summary": "Test Provider Config",
        "operationId": "test_provider_config_api_v1_provider_configs__provider_id__test_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionTestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ConnectionTestResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderConfigCreate": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Label"
          },
          "base_url": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Base Url"
          },
          "api_key": {
            "type": "string",
            "maxLength": 512,
            "minLength": 4,
            "title": "Api Key"
          },
          "default_model": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Default Model"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "label",
          "base_url",
          "api_key",
          "default_model"
        ],
        "title": "ProviderConfigCreate"
      },
      "ProviderConfigResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model"
          },
          "api_key_hint": {
            "type": "string",
            "title": "Api Key Hint"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled"
          },
          "last_test_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Test Status"
          },
          "last_test_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Test Error"
          },
          "last_tested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "base_url",
          "default_model",
          "api_key_hint",
          "is_enabled",
          "last_test_status",
          "last_test_error",
          "last_tested_at",
          "created_at",
          "updated_at"
        ],
        "title": "ProviderConfigResponse"
      },
      "ProviderConfigUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512,
                "minLength": 4
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "default_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          },
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled"
          }
        },
        "type": "object",
        "title": "ProviderConfigUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}